% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceCodeYuimaGUI.R
\name{yuimaGUI}
\alias{yuimaGUI}
\title{User-friendly Interface for the yuima package}
\usage{
yuimaGUI(theme = "black")
}
\arguments{
\item{theme}{GUI theme: "black" or "white".}
}
\value{
Starts yuima GUI
}
\description{
Runs yuima Graphical User Interface
}
\examples{
\dontrun{
yuimaGUI()
}

}
\author{
The YUIMA Project Team
}
