\name{yuimaGUI}
\alias{yuimaGUI}

\title{
User-friendly Interface for the yuima package
}
\description{
Runs yuima Graphical User Interface
}
\usage{
yuimaGUI(theme = "last")
}

\arguments{
  \item{theme}{GUI theme: black or white. By default the last one you used is chosen.}
}

\value{
Starts yuima GUI
}

\author{
The YUIMA Project Team
}

\examples{
\dontrun{
yuimaGUI()
}
}
